Move operators to session package
