+++ backend/setup.cfg
@@ -41,8 +41,8 @@ capellacollab.extensions.modelsources =
     git = t4cclient.extensions.modelsources.git
     t4c = t4cclient.extensions.modelsources.t4c
 capellacollab.operators =
-    k8s = t4cclient.core.operators.k8s:KubernetesOperator
-    docker = t4cclient.core.operators.docker:DockerOperator
+    k8s = t4cclient.sessions.operators.k8s:KubernetesOperator
+    docker = t4cclient.sessions.operators.docker:DockerOperator
 capellacollab.authentication.providers = 
     oauth = t4cclient.core.authentication.provider.oauth
     azure = t4cclient.core.authentication.provider.azure
