+++ backend/t4cclient/sessions/operators/docker.py
@@ -7,7 +7,7 @@ import shutil
 import typing as t
 
 from t4cclient.config import config
-from t4cclient.core.operators.abc import Operator
+from t4cclient.sessions.operators.abc import Operator
 
 import docker
