+++ backend/t4cclient/sessions/sessions.py
@@ -12,7 +12,7 @@ import requests
 
 # local:
 from t4cclient import config
-from t4cclient.core.operators import OPERATOR
+from t4cclient.sessions.operators import OPERATOR
 from t4cclient.sessions.schema import WorkspaceType
 from t4cclient.sessions.models import DatabaseSession
